♻️ Refactor :  complete 핸들러 Custom hook으로 분리하여 개선 및 불필요한 주석 제거
